+++ packages/client-node/__tests__/unit/node_config.test.ts
@@ -8,11 +8,11 @@ import { Buffer } from 'buffer'
 import http from 'http'
 import type { NodeClickHouseClientConfigOptions } from '../../src/config'
 import { NodeConfigImpl } from '../../src/config'
-import type {
-  CreateConnectionParams,
-  NodeBaseConnection,
+import {
+  type CreateConnectionParams,
+  type NodeBaseConnection,
+  NodeConnectionFactory,
 } from '../../src/connection'
-import * as c from '../../src/connection/create_connection'
 
 describe('[Node.js] Config implementation details', () => {
   describe('HandleImplSpecificURLParams', () => {
@@ -85,9 +85,10 @@ describe('[Node.js] Config implementation details', () => {
     let createConnectionStub: jasmine.Spy
     const fakeConnection = { test: true } as unknown as NodeBaseConnection
     beforeEach(() => {
-      createConnectionStub = spyOn(c, 'createConnection').and.returnValue(
-        fakeConnection,
-      )
+      createConnectionStub = spyOn(
+        NodeConnectionFactory,
+        'create',
+      ).and.returnValue(fakeConnection)
     })
 
     it('should create a connection with default KeepAlive settings', async () => {
@@ -104,6 +105,7 @@ describe('[Node.js] Config implementation details', () => {
         },
         http_agent: undefined,
         set_basic_auth_header: true,
+        capture_enhanced_stack_trace: false,
       } satisfies CreateConnectionParams)
       expect(createConnectionStub).toHaveBeenCalledTimes(1)
       expect(res).toEqual(fakeConnection)
@@ -129,6 +131,7 @@ describe('[Node.js] Config implementation details', () => {
         },
         http_agent: undefined,
         set_basic_auth_header: true,
+        capture_enhanced_stack_trace: false,
       } satisfies CreateConnectionParams)
       expect(createConnectionStub).toHaveBeenCalledTimes(1)
       expect(res).toEqual(fakeConnection)
@@ -158,6 +161,7 @@ describe('[Node.js] Config implementation details', () => {
         },
         http_agent: undefined,
         set_basic_auth_header: true,
+        capture_enhanced_stack_trace: false,
       } satisfies CreateConnectionParams)
       expect(createConnectionStub).toHaveBeenCalledTimes(1)
       expect(res).toEqual(fakeConnection)
@@ -187,6 +191,7 @@ describe('[Node.js] Config implementation details', () => {
         },
         http_agent: undefined,
         set_basic_auth_header: true,
+        capture_enhanced_stack_trace: false,
       } satisfies CreateConnectionParams)
       expect(createConnectionStub).toHaveBeenCalledTimes(1)
       expect(res).toEqual(fakeConnection)
@@ -215,6 +220,28 @@ describe('[Node.js] Config implementation details', () => {
         },
         http_agent: agent,
         set_basic_auth_header: false,
+        capture_enhanced_stack_trace: false,
+      } satisfies CreateConnectionParams)
+      expect(createConnectionStub).toHaveBeenCalledTimes(1)
+      expect(res).toEqual(fakeConnection)
+    })
+
+    it('should create a connection with enhanced stack traces option', async () => {
+      const nodeConfig: NodeClickHouseClientConfigOptions = {
+        url: new URL('https://localhost:8123'),
+        capture_enhanced_stack_trace: true,
+      }
+      const res = NodeConfigImpl.make_connection(nodeConfig as any, params)
+      expect(createConnectionStub).toHaveBeenCalledWith({
+        connection_params: params,
+        tls: undefined,
+        keep_alive: {
+          enabled: true,
+          idle_socket_ttl: 2500,
+        },
+        http_agent: undefined,
+        set_basic_auth_header: true,
+        capture_enhanced_stack_trace: true,
       } satisfies CreateConnectionParams)
       expect(createConnectionStub).toHaveBeenCalledTimes(1)
       expect(res).toEqual(fakeConnection)