+++ packages/client-node/__tests__/unit/node_create_connection.test.ts
@@ -2,14 +2,14 @@ import type { ConnectionParams } from '@clickhouse/client-common'
 import http from 'http'
 import https from 'node:https'
 import {
-  createConnection,
+  NodeConnectionFactory,
   type NodeConnectionParams,
   NodeHttpConnection,
   NodeHttpsConnection,
 } from '../../src/connection'
 import { NodeCustomAgentConnection } from '../../src/connection/node_custom_agent_connection'
 
-describe('[Node.js] createConnection', () => {
+describe('[Node.js] NodeConnectionFactory', () => {
   const keepAliveParams: NodeConnectionParams['keep_alive'] = {
     enabled: true,
     idle_socket_ttl: 2500,
@@ -26,18 +26,19 @@ describe('[Node.js] createConnection', () => {
   } as ConnectionParams
 
   it('should create an instance of HTTP adapter', async () => {
-    const adapter = createConnection({
+    const adapter = NodeConnectionFactory.create({
       connection_params: defaultConnectionParams,
       tls: tlsParams,
       keep_alive: keepAliveParams,
       http_agent: undefined,
       set_basic_auth_header: true,
+      capture_enhanced_stack_trace: false,
     })
     expect(adapter).toBeInstanceOf(NodeHttpConnection)
   })
 
   it('should create an instance of HTTPS adapter', async () => {
-    const adapter = createConnection({
+    const adapter = NodeConnectionFactory.create({
       connection_params: {
         ...defaultConnectionParams,
         url: new URL('https://localhost'),
@@ -46,13 +47,14 @@ describe('[Node.js] createConnection', () => {
       keep_alive: keepAliveParams,
       http_agent: undefined,
       set_basic_auth_header: true,
+      capture_enhanced_stack_trace: false,
     })
     expect(adapter).toBeInstanceOf(NodeHttpsConnection)
   })
 
   it('should throw if the supplied protocol is unknown', async () => {
     expect(() =>
-      createConnection({
+      NodeConnectionFactory.create({
         connection_params: {
           ...defaultConnectionParams,
           url: new URL('tcp://localhost'),
@@ -61,13 +63,14 @@ describe('[Node.js] createConnection', () => {
         keep_alive: keepAliveParams,
         http_agent: undefined,
         set_basic_auth_header: true,
+        capture_enhanced_stack_trace: false,
       }),
     ).toThrowError('Only HTTP and HTTPS protocols are supported')
   })
 
   describe('Custom HTTP agent', () => {
     it('should create an instance with a custom HTTP agent', async () => {
-      const adapter = createConnection({
+      const adapter = NodeConnectionFactory.create({
         connection_params: defaultConnectionParams,
         tls: tlsParams,
         keep_alive: keepAliveParams,
@@ -76,12 +79,13 @@ describe('[Node.js] createConnection', () => {
           maxSockets: 2,
         }),
         set_basic_auth_header: false,
+        capture_enhanced_stack_trace: false,
       })
       expect(adapter).toBeInstanceOf(NodeCustomAgentConnection)
     })
 
     it('should create an instance with a custom HTTPS agent', async () => {
-      const adapter = createConnection({
+      const adapter = NodeConnectionFactory.create({
         connection_params: defaultConnectionParams,
         tls: tlsParams,
         keep_alive: keepAliveParams,
@@ -90,6 +94,7 @@ describe('[Node.js] createConnection', () => {
           maxSockets: 2,
         }),
         set_basic_auth_header: true,
+        capture_enhanced_stack_trace: false,
       })
       expect(adapter).toBeInstanceOf(NodeCustomAgentConnection)
     })