+++ packages/client-node/src/connection/create_connection.ts
@@ -15,38 +15,46 @@ export interface CreateConnectionParams {
   keep_alive: NodeConnectionParams['keep_alive']
   http_agent: http.Agent | https.Agent | undefined
   set_basic_auth_header: boolean
+  capture_enhanced_stack_trace: boolean
 }
 
-export function createConnection({
-  connection_params,
-  tls,
-  keep_alive,
-  http_agent,
-  set_basic_auth_header,
-}: CreateConnectionParams): NodeBaseConnection {
-  if (http_agent !== undefined) {
-    return new NodeCustomAgentConnection({
-      ...connection_params,
-      set_basic_auth_header,
-      keep_alive, // only used to enforce proper KeepAlive headers
-      http_agent,
-    })
-  }
-  switch (connection_params.url.protocol) {
-    case 'http:':
-      return new NodeHttpConnection({
+/** A factory for easier mocking after Node.js 22.18 */
+export class NodeConnectionFactory {
+  static create({
+    connection_params,
+    tls,
+    keep_alive,
+    http_agent,
+    set_basic_auth_header,
+    capture_enhanced_stack_trace,
+  }: CreateConnectionParams): NodeBaseConnection {
+    if (http_agent !== undefined) {
+      return new NodeCustomAgentConnection({
         ...connection_params,
         set_basic_auth_header,
-        keep_alive,
+        capture_enhanced_stack_trace,
+        keep_alive, // only used to enforce proper KeepAlive headers
+        http_agent,
       })
-    case 'https:':
-      return new NodeHttpsConnection({
-        ...connection_params,
-        set_basic_auth_header,
-        keep_alive,
-        tls,
-      })
-    default:
-      throw new Error('Only HTTP and HTTPS protocols are supported')
+    }
+    switch (connection_params.url.protocol) {
+      case 'http:':
+        return new NodeHttpConnection({
+          ...connection_params,
+          set_basic_auth_header,
+          capture_enhanced_stack_trace,
+          keep_alive,
+        })
+      case 'https:':
+        return new NodeHttpsConnection({
+          ...connection_params,
+          set_basic_auth_header,
+          capture_enhanced_stack_trace,
+          keep_alive,
+          tls,
+        })
+      default:
+        throw new Error('Only HTTP and HTTPS protocols are supported')
+    }
   }
 }