+++ CHANGELOG.md
@@ -1,3 +1,49 @@
+# 1.12.0
+
+## Types
+
+- Add missing `allow_experimental_join_condition` to `ClickHouseSettings` typing. ([#430], [looskie])
+- Fixed `JSONEachRowWithProgress` TypeScript flow after the breaking changes in [ClickHouse 25.1]. `RowOrProgress<T>` now has an additional variant: `SpecialEventRow<T>`. The library now additionally exports the `parseError` method, and newly added `isRow` / `isException` type guards. See the updated [JSONEachRowWithProgress example] ([#443])
+- Added missing `allow_experimental_variant_type` (24.1+), `allow_experimental_dynamic_type` (24.5+), `allow_experimental_json_type` (24.8+), `enable_json_type` (25.3+), `enable_time_time64_type` (25.6+) to `ClickHouseSettings` typing. ([#445])
+
+## Improvements
+
+- Add a warning on a socket closed without fully consuming the stream (e.g., when using `query` or `exec` method). ([#441])
+- (Node.js only) An option to use a simple SELECT query for ping checks instead of `/ping` endpoint. See the new optional argument to the `ClickHouseClient.ping` method and `PingParams` typings. Note that the Web version always used a SELECT query by default, as the `/ping` endpoint does not support CORS, and that cannot be changed. ([#442])
+
+## Other
+
+- The project now uses [Codecov] instead of SonarCloud for code coverage reports. ([#444])
+
+[#430]: https://github.com/ClickHouse/clickhouse-js/pull/430
+[#441]: https://github.com/ClickHouse/clickhouse-js/pull/441
+[#442]: https://github.com/ClickHouse/clickhouse-js/pull/442
+[#443]: https://github.com/ClickHouse/clickhouse-js/pull/443
+[#444]: https://github.com/ClickHouse/clickhouse-js/pull/444
+[#445]: https://github.com/ClickHouse/clickhouse-js/pull/445
+[looskie]: https://github.com/looskie
+[ClickHouse 25.1]: https://github.com/ClickHouse/ClickHouse/pull/74181
+[JSONEachRowWithProgress example]: https://github.com/ClickHouse/clickhouse-js/blob/main/examples/node/select_json_each_row_with_progress.ts
+[Codecov]: https://codecov.io/gh/ClickHouse/clickhouse-js
+
+# 1.11.2 (Common, Node.js)
+
+A minor release to allow further investigation regarding uncaught error issues with [#410].
+
+## Types
+
+- Added missing `lightweight_deletes_sync` typing to `ClickHouseSettings` ([#422], [pratimapatel2008])
+
+## Improvements (Node.js)
+
+- Added a new configuration option: `capture_enhanced_stack_trace`; see the JS doc in the Node.js client package. Note that it is disabled by default due to a possible performance impact. ([#427])
+- Added more try-catch blocks to the Node.js connection layer. ([#427])
+
+[#410]: https://github.com/ClickHouse/clickhouse-js/pull/410
+[#422]: https://github.com/ClickHouse/clickhouse-js/pull/422
+[#427]: https://github.com/ClickHouse/clickhouse-js/pull/427
+[pratimapatel2008]: https://github.com/pratimapatel2008
+
 # 1.11.1 (Common, Node.js, Web)
 
 ## Bug fixes
@@ -128,7 +174,6 @@
   While the original intention was to use this function internally for `Native`/`RowBinaryWithNamesAndTypes` data formats headers parsing, it can be useful for other purposes as well (e.g., interfaces generation, or custom JSON serializers).
 
   NB: currently unsupported source types to parse:
-
   - Geo
   - (Simple)AggregateFunction
   - Nested
@@ -188,7 +233,6 @@
 ## Improvements
 
 - Re-exported several constants from the `@clickhouse/client-common` package for convenience:
-
   - `SupportedJSONFormats`
   - `SupportedRawFormats`
   - `StreamableFormats`
@@ -497,7 +541,6 @@ This release primarily focuses on improving the Keep-Alive mechanism's reliabili
 - Idle sockets timeout rework; now, the client attaches internal timers to idling sockets, and forcefully removes them from the pool if it considers that a particular socket is idling for too long. The intention of this additional sockets housekeeping is to eliminate "Socket hang-up" errors that could previously still occur on certain configurations. Now, the client does not rely on KeepAlive agent when it comes to removing the idling sockets; in most cases, the server will not close the socket before the client does.
 - There is a new `keep_alive.idle_socket_ttl` configuration parameter. The default value is `2500` (milliseconds), which is considered to be safe, as [ClickHouse versions prior to 23.11 had `keep_alive_timeout` set to 3 seconds by default](https://github.com/ClickHouse/ClickHouse/commit/1685cdcb89fe110b45497c7ff27ce73cc03e82d1), and `keep_alive.idle_socket_ttl` is supposed to be slightly less than that to allow the client to remove the sockets that are about to expire before the server does so.
 - Logging improvements: more internal logs on failing requests; all client methods except ping will log an error on failure now. A failed ping will log a warning, since the underlying error is returned as a part of its result. Client logging still needs to be enabled explicitly by specifying the desired `log.level` config option, as the log level is `OFF` by default. Currently, the client logs the following events, depending on the selected `log.level` value:
-
   - `TRACE` - low-level information about the Keep-Alive sockets lifecycle.
   - `DEBUG` - response information (without authorization headers and host info).
   - `INFO` - still mostly unused, will print the current log level when the client is initialized.